Refactor server key validation across services and update dependencies
