+++ crates/services/fitness/src/template_management.rs
@@ -13,7 +13,7 @@ use dependent_models::{
 };
 use dependent_utils::{
     expire_user_workout_templates_list_cache, get_focused_workout_summary_with_exercises,
-    user_workout_templates_list as get_user_workout_templates_list,
+    is_server_key_validated, user_workout_templates_list as get_user_workout_templates_list,
 };
 use fitness_models::{
     ProcessedExercise, UserWorkoutInput, WorkoutInformation, WorkoutSetRecord, WorkoutSummary,
@@ -46,7 +46,7 @@ pub async fn create_or_update_user_workout_template(
     user_id: String,
     input: UserWorkoutInput,
 ) -> Result<String> {
-    server_key_validation_guard(ss.is_server_key_validated().await?).await?;
+    server_key_validation_guard(is_server_key_validated(ss).await?).await?;
     let mut summary = WorkoutSummary::default();
     let mut information = WorkoutInformation {
         comment: input.comment,
@@ -136,7 +136,7 @@ pub async fn delete_user_workout_template(
     user_id: String,
     workout_template_id: String,
 ) -> Result<bool> {
-    server_key_validation_guard(ss.is_server_key_validated().await?).await?;
+    server_key_validation_guard(is_server_key_validated(ss).await?).await?;
     let Some(wkt) = WorkoutTemplate::find_by_id(workout_template_id)
         .filter(workout_template::Column::UserId.eq(&user_id))
         .one(&ss.db)
