+++ crates/utils/dependent/src/lib.rs
@@ -1,6 +1,9 @@
 mod analytics_operations;
 pub use analytics_operations::*;
 
+mod core_operations;
+pub use core_operations::*;
+
 mod provider_services;
 pub use provider_services::*;
